support signing empty param values, e.g. 's.q='
As long as an empty value for a param (e.g., s.q) is legal,
we should support signing it. Current implementation simply
drops params with empty values, which without reporting back
to the caller, results in invalid auth headers. This should
also be supported because in practice, sometimes having a parameter
key present with an empty value can differ semantically from
having no value explicitly specified for that parameter.